+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
 ## Changed
+- fixed link in mention notifications
 - title for new file
 
 ## 7.3.0